2023/3/7 Jp logic排查，套餐中查看菜品问题，套餐分类中查询问题解决

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,11 +2,15 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -21,7 +25,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     CategoryMapper categoryMapper;
-
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
 
     @Override
@@ -40,11 +47,9 @@ public class CategoryServiceImpl implements CategoryService {
 
         PageHelper.startPage(categoryPageQueryDTO.getPage(),categoryPageQueryDTO.getPageSize());
 
-        Page<Category> result = categoryMapper.queryCategoryPage(categoryPageQueryDTO);
-        long total = result.getTotal();
-        List<Category> records = result.getResult();
+        Page<Category> page = categoryMapper.queryCategoryPage(categoryPageQueryDTO);
 
-        return new PageResult(total,records);
+        return new PageResult(page.getTotal(),page.getResult());
     }
 
     @Override
@@ -65,8 +70,25 @@ public class CategoryServiceImpl implements CategoryService {
         categoryMapper.update(category);
     }
 
-    @Override
+    /**
+     * todo 分类：与菜品套餐存在关联
+     * @param id
+     */
     public void deleteCategory1(Long id) {
+        //查询当前分类是否关联了菜品，如果关联了就抛出业务异常
+        Integer count = dishMapper.countByCategoryId(id);
+        if(count > 0){
+            //当前分类下有菜品，不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+
+        //查询当前分类是否关联了套餐，如果关联了就抛出业务异常
+        count = setmealMapper.countByCategoryId(id);
+        if(count > 0){
+            //当前分类下有菜品，不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+
         categoryMapper.delete(id);
     }
